Refactor duplicate MatchAnyParser and MatchAllParser code into MatchParser
